+++ packages/cargo/jns42-core/src/names/mod.rs
@@ -1,5 +1,12 @@
+mod names;
+mod names_builder;
+mod part;
 mod sentence;
 mod word;
 
+use part::*;
+use word::*;
+
+pub use names::*;
+pub use names_builder::*;
 pub use sentence::*;
-pub use word::*;
