+++ packages/cargo/jns42-core/src/utils/mod.rs
@@ -1,5 +1,4 @@
 pub mod json_pointer;
 pub mod merge;
-pub mod names;
 pub mod product;
 pub mod url;
